% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFitCompare.R
\name{lmFitComp}
\alias{lmFitComp}
\title{lmFitComp}
\usage{
lmFitComp(beta, pheno)
}
\arguments{
\item{beta}{A matrix of beta-values between 0 and 1 with columns as samples and rows as probes}

\item{pheno}{A phenotype data frame containing variables of interest on samples. Must be ordered the same as beta matrix.}
}
\value{
number of hypomethylated sites (down), insignificant sites, and hypermethylated sites (up)
}
\description{
Takes a beta matrix with columns as samples, probes as rows
and a phenotype dataframe with samples as rows and features as columns
}
\examples{
lmFitComp(beta_matrix, phenotype_dataframe)
}
